Functionality Implementation
I was able to implement the functionality successfully. A prompt asks the user for an inquiry before generating the first fortuine, then the fortune is generated:  3 cards are displayed and the user is able to request and additional fortune.
Card and Deck objects were altered slightly.
Not yet implemented:
- loading the images
- loading/showing the description
-saving the fortune to a file.

Testing Needed
-user prompt: how to prevent from continuing to generate fortune if the text field is empty.
-future functionalities.